break down NDEx dialog into smaller components for better maintainability
